MADTwin/layout/pyproject.toml
[tool.poetry]
name = "uwbmap"
version = "0.1.0"
description = ""
authors = ["Ebert Schoofs <ebertschoofs@gmail.com>"]

[tool.poetry.scripts]
ampyfier = "layout.uwbmap.__main__:__main__"

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
qtwidgets = "^0.18"
paho-mqtt = "^1.6.1"
py4j = "^0.10.9"
PyQt6 = "^6.2.3"
matplotlib = "^3.6.1"
similaritymeasures = "^0.6.0"
PySide6 = "^6.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

